feat: implement custom BuildPackStore initialization with suffix
Running multiple integration test suits in parallel does not
work as is with the BuildpackStore class. The first suit that
creates the store "wins" and the next cannot open it.
Also just changing the cacheManager to make it point
somewhere else does not work and the suite ends with "the cache
manager is not loaded properly".

This helper re-implements the original method but adds a
suffix parameter so that each test can have it's own store.