feat: initial implementation of pipenv
pipenv comes from pipenv-install

The generated plan as been refactored (for pip as well) to
provide and require an new entry that allows to differentiate
which shall be used at build time since both pip-install and
pipenv-install provide site-packages.
The provides name could have been changed however for backward
compatibility, this solution has been select. This can be
changed in a followup patch in sync with a change to python-start.